backend/requirements.txt

# CORE DJANGO

Django==5.1.0
asgiref==3.8.1
sqlparse==0.5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
django-cors-headers==4.4.0

# PRODUCTION DEPLOYMENT
gunicorn==23.0.0
whitenoise==6.7.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1


# TENSORFLOW 2.13 (ALL COMPATIBLE VERSIONS)

tensorflow-cpu==2.13.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
keras==2.13.1

# TensorFlow Dependencies
absl-py==1.4.0
astunparse==1.6.3
flatbuffers==23.5.26
gast==0.4.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
libclang==16.0.0
opt-einsum==3.3.0
protobuf==4.23.4
wrapt==1.14.1
typing-extensions==4.5.0
termcolor==2.3.0


# GOOGLE AUTH

google-auth==2.17.3
google-auth-oauthlib==1.0.0
cachetools==5.3.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
rsa==4.9
oauthlib==3.2.2
requests-oauthlib==1.3.1
numpy==1.24.3
packaging==23.1
opencv-python-headless==4.8.0.74
Pillow==10.0.0


# WEB SCRAPING

beautifulsoup4==4.12.3
soupsieve==2.5


# TRANSLATION (Alternative to googletrans)

deep-translator==1.11.4


# HTTP CLIENTS

requests==2.31.0
urllib3==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4

# UTILITIES
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.3.7
Markdown==3.4.4
MarkupSafe==2.1.3